added batch renderer

+++ Engine/Platforms/Vulkan/Impl/Vk1Device.h
@@ -204,7 +204,7 @@ namespace PlatformVK
 		ModulePtr				GetCommandBuilder ()		const	{ return _commandBuilder; }
 		ModulePtr				GetCurrentFramebuffer ()	const	{ return _framebuffers[ _currentImageIndex ]; }
 		uint					GetImageIndex ()			const	{ return _currentImageIndex; }
-		uint					GetSwapchainLength ()		const	{ return _framebuffers.Count(); }
+		uint					GetSwapchainLength ()		const	{ return (uint)_framebuffers.Count(); }
 
 		uint2 const&			GetSurfaceSize ()			const	{ return _surfaceSize; }
